Implemented Server connection and message sending stuff from Tutorial week 2
Added the option to connect to server and option to send messages to server and see them back, need to understand that code and add comments, fix the Design of GUI because my double click mousepad messed everyhthing up there. Also added the names of JTextAreas for now and made Send button to appear after connect, and made sure the Input JTextArea becomes editable after the connection has made.
Things to do with the stuff added (Not full list, just for now):
Major things:
1)Need to make that many people can connect, because now it is that server is created and thats all and u send info there. (Connect to the IP address that are created by the Host and make sure others can connect throught the Connect screen. )
Minor bits:
1)Fix layout of the GUI.
2) Make txt file for connection info then connection Info box needs to be writed into after connection from txt file.
3) Need to make People online lister and fill in the Onliners JTextArea on the right.