Added AdminPanel.
Admin Panel got added to start the server! (Needs to be updated with Inputs with Username and password to start the server)
Added Catch for connection to Offline server, deleted Connect button disabling due to if the connection failed user cant connect anymore. Need to make sure the messages pops up in MainScreen as for now they only get sended to the AdminPanel.