Made combobox w users and private chat
Made combobox that has online users in it and private chat that the program knows what is selected in combobox and makes a connection with the IP address of that person.
When barnabe adds the Online ips and connections then will add that people gonna send their ip that will just replace the ips to connect to when doing private chat.